% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donorheuristics-package.R
\docType{package}
\name{donorheuristics-package}
\alias{donorheuristics-package}
\alias{donorheuristics}
\title{The 'donorheuristics' package.}
\description{
Research on how donors respond to organizational and structural
  NGO characteristics
}
